backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aniso8601==10.0.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
executing==2.1.0
filelock==3.16.1
finnhub-python==2.4.22
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==2.3.0
flask-restx==1.3.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langsmith==0.2.10
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
openai==1.59.6
orjson==3.10.14
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
sympy==1.13.1
ta==0.11.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarl==1.18.3
yfinance==0.2.51
youtube-transcript-api==0.6.3
